Refactor: Add package structure
패키지 구조 추가
